mprove navigation bar with theme toggle and fixes
Added theme and language switch button to the navigation bar
Fixed colors to properly support light theme
Added blur effect when content goes under the navigation bar
Fixed issue where scrollbar disappeared when opening a menu
