+++ frontend/vizzy/app/page.tsx
@@ -1,44 +1,16 @@
-'use client';
-
-import { LanguageSwitcher } from '@/components/i18n/language-switcher';
 import { Footer } from '@/components/layout/footer';
-import { ThemeToggle } from '@/components/theme/theme-toggle';
-import { createClient } from '@/utils/supabase/client';
+import NavBar from '@/components/layout/nav-bar';
 import { useTranslations } from 'next-intl';
 import Image from 'next/image';
 
 export default function Home() {
-  const getAccessToken = async () => {
-    const supabase = await createClient();
-    const session = await supabase.auth.getSession();
-    const accessToken = session.data.session?.access_token;
-    console.log('Access Token:', accessToken);
-  };
-
   const t = useTranslations('Testing');
-
-  // Fetch cookies from the backend
-  const setCookies = async () => {
-    try {
-      const response = await fetch('https://localhost:3000/set-cookie', {
-        method: 'GET',
-        credentials: 'include',
-      });
-
-      const data = await response.text();
-      console.log(data);
-    } catch (error) {
-      console.error('Error setting cookie:', error);
-    }
-  };
 
   return (
     <div>
-      <LanguageSwitcher />
-      <h1>{t('Test')}</h1>
+      <NavBar userName={''} avatarUrl={'https://github.com/shadcn.png'} />
       <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
         <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-          <ThemeToggle />
           <Image
             className="dark:invert"
             src="/next.svg"
@@ -57,6 +29,8 @@ export default function Home() {
             </li>
             <li>Save and see your changes instantly.</li>
           </ol>
+
+          <h1>{t('Test')}</h1>
 
           <div className="flex gap-4 items-center flex-col sm:flex-row">
             <a
@@ -83,22 +57,6 @@ export default function Home() {
               Read our docs
             </a>
           </div>
-
-          {/* Button to fetch and log the access token */}
-          <button
-            onClick={getAccessToken}
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-blue-500 text-white gap-2 hover:bg-blue-600 text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 mt-4"
-          >
-            Get Access Token
-          </button>
-
-          {/* Button to set cookies */}
-          <button
-            onClick={setCookies}
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-green-500 text-white gap-2 hover:bg-green-600 text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 mt-4"
-          >
-            Set Cookie
-          </button>
         </main>
         <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
           <a
