+++ frontend/vizzy/components/layout/nav-bar.tsx
@@ -1,0 +1,46 @@
+import Logo from '@/components/logo';
+import { UserMenu } from '@/components/layout/user-menu';
+import { JSX } from 'react';
+import { LanguageSwitcher } from '../i18n/language-switcher';
+import { ThemeToggle } from '../theme/theme-toggle';
+import Link from 'next/link';
+
+interface NavBarProps {
+  userName: string;
+  avatarUrl: string;
+}
+
+/**
+ * Navigation bar component that displays the application logo, user avatar,
+ * and a dropdown menu with various options.
+ *
+ * This component is designed to be simple and flexible:
+ * - The logo is displayed on the left.
+ * - The user's avatar is shown on the right, which opens a dropdown menu when clicked.
+ *
+ * @component
+ * @param {NavBarProps} props - The component's properties.
+ * @param {string} props.userName - The name of the user, displayed in the dropdown menu and used for the avatar's fallback text.
+ * @param {string} props.avatarUrl - The URL of the user's avatar image.
+ * @returns {JSX.Element} The rendered navigation bar component.
+ */
+export default function NavBar({
+  userName,
+  avatarUrl,
+}: NavBarProps): JSX.Element {
+  return (
+    <div className="fixed top-0 left-0 right-0 text-white h-16 flex items-center justify-between px-6 bg-background/80 backdrop-blur-sm z-50 border-b">
+      <div className="flex items-center">
+        <Link className="cursor-pointer" href="/">
+          <Logo />
+        </Link>
+      </div>
+
+      <div className="flex items-center gap-2">
+        <ThemeToggle />
+        <LanguageSwitcher />
+        <UserMenu userName={userName} avatarUrl={avatarUrl} />
+      </div>
+    </div>
+  );
+}
