--- frontend/vizzy/components/ui/nav-bar.tsx
@@ -1,92 +1,0 @@
-"use client"
-
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
-import { User, Store, CreditCard, Settings, LogOut } from "lucide-react"
-import Logo from "@/components/logo"
-
-interface NavBarProps {
-  userName: string
-  avatarUrl: string
-}
-
-/**
- * Navigation bar component that displays the application logo, user avatar, 
- * and a dropdown menu with various options.
- *
- * This component is designed to be simple and flexible:
- * - The logo is displayed on the left.
- * - The user's avatar is shown on the right, which opens a dropdown menu when clicked.
- * - The dropdown menu contains options for Profile, Proposals, Transactions, Settings, and Logout.
- *
- * @component
- * @param {NavBarProps} props - The component's properties.
- * @param {string} props.userName - The name of the user, displayed in the dropdown menu and used for the avatar's fallback text.
- * @param {string} props.avatarUrl - The URL of the user's avatar image.
- * @returns {JSX.Element} The rendered navigation bar component.
- */
-
-export default function NavBar({ userName, avatarUrl }: NavBarProps) {
-  return (
-    <div className="bg-zinc-950 min-h-screen">
-      <div className="relative bg-black text-white h-16 flex items-center justify-between px-6">
-        <div className="flex items-center">
-          <Logo />
-        </div>
-
-        <div className="flex items-center gap-4">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Avatar className="cursor-pointer">
-                <AvatarImage src={avatarUrl} alt={userName} />
-                <AvatarFallback>
-                  {userName
-                    .split(" ")
-                    .map((n) => n[0])
-                    .join("")
-                    .toUpperCase()}
-                </AvatarFallback>
-              </Avatar>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent className="w-64 bg-zinc-900 border-zinc-800" align="end">
-              <DropdownMenuLabel>{userName}</DropdownMenuLabel>
-              <DropdownMenuSeparator />
-              <DropdownMenuGroup>
-                <DropdownMenuItem>
-                  <User className="mr-2 h-4 w-4" />
-                  <span>Perfil</span>
-                </DropdownMenuItem>
-                <DropdownMenuItem>
-                  <Store className="mr-2 h-4 w-4" />
-                  <span>Propostas</span>
-                </DropdownMenuItem>
-                <DropdownMenuItem>
-                  <CreditCard className="mr-2 h-4 w-4" />
-                  <span>Transações</span>
-                </DropdownMenuItem>
-                <DropdownMenuItem>
-                  <Settings className="mr-2 h-4 w-4" />
-                  <span>Definições</span>
-                </DropdownMenuItem>
-              </DropdownMenuGroup>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem>
-                <LogOut className="mr-2 h-4 w-4" />
-                <span>Terminar Sessão</span>
-              </DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </div>
-      </div>
-    </div>
-  )
-}
-
